# apps/taptrade-platform/go-platform/services/gateway/api/openapi.yaml
openapi: 3.0.3
info:
  title: Taya NA Predict API
  version: "1.0.0"
  description: >
    TapTrade prediction-market gateway API for non-redeemable gameplay points.
    Binary YES/NO outcomes use point-cents subunits on a 0-100 implied
    probability scale; settled winning shares return 100 point-cents per share.
    Points are gameplay-only units with no external value path.

    Coverage: this spec documents the core public, trading, portfolio,
    responsible-play, reward, loyalty, social, leaderboard, bot, auth proxy,
    admin account-review, admin discovery taxonomy, admin market operations,
    and admin market-lifecycle surfaces verified against the live handlers and
    DTOs as of 2026-06. CMS and remaining compatibility surfaces are
    intentionally out of scope here and tracked for documentation in
    IMPROVEMENT_PLAN.md P2-05 / G-04. The previous inherited spec is archived
    outside this launch document.
servers:
  - url: http://localhost:18080
    description: Local development
components:
  securitySchemes:
    SessionCookie:
      type: apiKey
      in: cookie
      name: access_token
      description: HttpOnly session cookie issued by the auth service. Mutating requests also require the X-CSRF-Token header (double-submit cookie).
    BotApiKey:
      type: http
      scheme: bearer
      description: "Bot/partner API key: Authorization: Bearer tna_<prefix>_<secret>. Scopes: read, trade."
  schemas:
...
    Dispute:
      type: object
      description: Holder challenge against a proposed resolution.
      properties:
        id: { type: string }
        marketId: { type: string }
        userId: { type: string }
        reason: { type: string }
        status: { type: string, enum: [open, upheld, rejected, withdrawn] }
        resolutionNote: { type: string, nullable: true }
        bondPoints: { type: integer, format: int64, description: Reserved challenge point-cents, currently zero for launch. }
        unit: { type: string, enum: [PTS] }
        createdAt: { type: string, format: date-time }
        resolvedAt: { type: string, format: date-time, nullable: true }
        resolvedBy: { type: string, nullable: true }
    DisputeCreateRequest:
      type: object
      properties:
        marketId: { type: string }
        reason: { type: string }
      required: [marketId, reason]
...
paths:
...
  /api/v1/disputes:
    get:
      summary: List market disputes
      description: Authenticated holders can review disputes for a market while it is in a proposed-resolution flow.
      security: [{ SessionCookie: [] }]
      parameters:
        - { name: marketId, in: query, required: true, schema: { type: string } }
      responses:
        "200":
          description: Dispute list
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items: { $ref: "#/components/schemas/Dispute" }
        "400": { description: marketId is required }
        "401": { description: Authentication required }
    post:
      summary: File a market dispute
      description: Allows an authenticated position holder to challenge a proposed resolution. Filing a dispute blocks finalization until admin review and does not move points.
      security: [{ SessionCookie: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/DisputeCreateRequest" }
      responses:
        "201":
          description: Created dispute
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Dispute" }
        "400": { description: Invalid dispute request }
        "401": { description: Authentication required }
        "403": { description: Only position holders may dispute }
        "409": { description: Market is not in a disputable state }
        "429": { description: Too many dispute submissions }
...